requirements.txt
snowflake-connector-python
google-cloud-storage
pandas
pyarrow
python-dotenv
toml
